# go-server/internal/gen/server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stick-It RESP API
  version: 1.0.1
servers:
- url: https://stick-it.lr-projects.de
- url: http://localhost:8081
...
paths:
...
  /api/v2/users/{userId}:
    delete:
      operationId: deleteUser
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/code"
      responses:
        "200":
          description: Deleted
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: User does not exist
      security:
      - token: []
      summary: Delete a user by userId
      tags:
      - users
    get:
      operationId: getUser
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userInfoDto"
          description: Get successfull
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: User does not exist
      security:
      - token: []
      summary: Get a user by userId
      tags:
      - users
    put:
      operationId: updateUser
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userUpdateDto"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userUpdateResponseDto"
          description: Update successfull
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: User does not exist
      security:
      - token: []
      summary: Update user information by userId
      tags:
      - users
  /api/v2/users/{userId}/profile_picture_small:
    get:
      operationId: getUserProfileImageSmall
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      - description: "When true, this endpoint redirects directly to the target image\
          \ otherwise the image URL is returned"
        in: query
        name: redirect
        required: false
        schema:
          default: false
          type: boolean
      responses:
        "200":
          content:
            image/*:
              schema:
                $ref: "#/components/schemas/imageUrl"
          description: Successful response
        "301":
          description: Successful redirect to image
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: User does not exist
      security:
      - token: []
      summary: Get the small profile picture of a user by userId
      tags:
      - users
  /api/v2/users/{userId}/profile_picture:
    get:
      operationId: getUserProfileImage
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      - description: "When true, this endpoint redirects directly to the target image\
          \ otherwise the image URL is returned"
        in: query
        name: redirect
        required: false
        schema:
          default: false
          type: boolean
      responses:
        "200":
          content:
            image/*:
              schema:
                $ref: "#/components/schemas/imageUrl"
          description: Successful response
        "301":
          description: Successful redirect to image
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: User does not exist
      security:
      - token: []
      summary: Get the profile picture of a user by userId
      tags:
      - users
  /api/v2/users/{userId}/xp:
    get:
      operationId: getUserXp
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userXpDto"
          description: Successfully returned user's xp
        "404":
          description: User not found
        "403":
          description: Token not of requested user
        "401":
          description: Unauthorized
        "400":
          description: Bad request parameter format
      security:
      - token: []
      summary: Get user's xp
      tags:
      - users
  /api/v2/users/{userId}/achievements:
    get:
      operationId: getUserAchievements
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userAchievementsDto"
          description: Successfully returned user's achievements
        "404":
          description: User not found
        "403":
          description: Token not of requested user
      security:
      - token: []
      summary: Get user's achievements
      tags:
      - users
  /api/v2/users/{userId}/achievements/{achievementId}:
    post:
      operationId: claimUserAchievement
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      - in: path
        name: achievementId
        required: true
        schema:
          minimum: 0
          type: integer
      responses:
        "200":
          description: Successfully claimed achievement
        "404":
          description: User not found
        "403":
          description: Token not of requested user
        "400":
          description: Not fulfilled conditions
      security:
      - token: []
      summary: Claim an achievement
      tags:
      - users
...
components:
  schemas:
    id:
      format: uuid
      type: string
...
    longString:
      maxLength: 500
      type: string
...
    imageUrl:
      example: https://minio.lr-projects/stick-it/1/1.png
      format: url
      type: string
...
    tokenResponseDto:
      example:
        accessToken: accessToken
        userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        refreshToken: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        refreshToken:
          format: uuid
          type: string
        accessToken:
          maxLength: 500
          type: string
        userId:
          format: uuid
          type: string
      required:
      - accessToken
      - refreshToken
      - userId
      type: object
...
    userInfoDto:
      example:
        bestSeason:
          season:
            month: 6
            year: 1
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            seasonNumber: 5
          rank: 2
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          points: 5
        description: description
        isMessagingRegistered: true
        userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        username: username
        selectedBatch: 0
      properties:
        username:
          maxLength: 256
          type: string
        userId:
          format: uuid
          type: string
        description:
          maxLength: 500
          type: string
        selectedBatch:
          nullable: true
          type: integer
        bestSeason:
          $ref: "#/components/schemas/seasonItemDto"
        isMessagingRegistered:
          description: Flag for whether the user registered to receive push messages.
            Only set for the current user.
          nullable: true
          type: boolean
      required:
      - userId
      - username
      type: object
    userUpdateDto:
      properties:
        email:
          format: email
          maximum: 265
          type: string
        password:
          maxLength: 256
          type: string
        image:
          example: iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNk+A8AAQUBAScY42YAAAAASUVORK5CYII=
          format: byte
          type: string
        username:
          maxLength: 256
          type: string
        description:
          maxLength: 500
          type: string
        selectedBatch:
          nullable: true
          type: integer
        messagingToken:
          maxLength: 256
          type: string
      type: object
    userUpdateResponseDto:
      example:
        userInfoDto:
          bestSeason:
            season:
              month: 6
              year: 1
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              seasonNumber: 5
            rank: 2
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            points: 5
          description: description
          isMessagingRegistered: true
          userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          username: username
          selectedBatch: 0
        profileImageSmall: https://minio.lr-projects/stick-it/1/1.png
        profileImage: https://minio.lr-projects/stick-it/1/1.png
        userTokenDto:
          accessToken: accessToken
          userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          refreshToken: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        userTokenDto:
          $ref: "#/components/schemas/tokenResponseDto"
        profileImageSmall:
          example: https://minio.lr-projects/stick-it/1/1.png
          format: url
          type: string
        profileImage:
          example: https://minio.lr-projects/stick-it/1/1.png
          format: url
          type: string
        userInfoDto:
          $ref: "#/components/schemas/userInfoDto"
      type: object
    code:
      maxLength: 6
      minLength: 6
      type: integer
    userXpDto:
      example:
        currentLevel: 0
        totalXp: 0
        nextLevelXp: 0
        currentLevelXp: 0
      properties:
        totalXp:
          minimum: 0
          type: integer
        currentLevel:
          minimum: 0
          type: integer
        currentLevelXp:
          minimum: 0
          type: integer
        nextLevelXp:
          minimum: 0
          type: integer
      required:
      - currentLevel
      - currentLevelXp
      - nextLevelXp
      - totalXp
      type: object
    userAchievementsDto:
      items:
        $ref: "#/components/schemas/userAchievementsDto_inner"
      type: array
...
    seasonItemDto:
      example:
        season:
          month: 6
          year: 1
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          seasonNumber: 5
        rank: 2
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        points: 5
      properties:
        id:
          format: uuid
          type: string
        season:
          $ref: "#/components/schemas/seasonDto"
        points:
          description: Number of sticks
          type: integer
        rank:
          description: Rank reached at the end of the season
          type: integer
      required:
      - id
      - points
      - rank
      - season
      type: object
...
    seasonDto:
      example:
        month: 6
        year: 1
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        seasonNumber: 5
      properties:
        id:
          format: uuid
          type: string
        month:
          type: integer
        year:
          type: integer
        seasonNumber:
          type: integer
      required:
      - id
      - month
      - seasonNumber
      - year
      type: object
...
    userAchievementsDto_inner:
      example:
        claimed: true
        achievementId: 0
        thresholdValue: 6
        thresholdUp: true
        currentValue: 1
      properties:
        achievementId:
          type: integer
        claimed:
          type: boolean
        thresholdValue:
          type: integer
        currentValue:
          type: integer
        thresholdUp:
          type: boolean
      required:
      - achievementId
      - claimed
      - currentValue
      - thresholdUp
      - thresholdValue
      type: object
...
  securitySchemes:
    token:
      bearerFormat: JWT
      scheme: bearer
      type: http