# go-server/internal/gen/server/api_users.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Stick-It RESP API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.1
 */

package genserver

import (
	"encoding/json"
	"errors"
	"io"
	"net/http"
	"strings"

	"github.com/go-chi/chi/v5"
)

// UsersAPIController binds http requests to an api service and writes the service results to the http response
type UsersAPIController struct {
	service      UsersAPIServicer
	errorHandler ErrorHandler
}

// UsersAPIOption for how the controller is set up.
type UsersAPIOption func(*UsersAPIController)

// WithUsersAPIErrorHandler inject ErrorHandler into controller
func WithUsersAPIErrorHandler(h ErrorHandler) UsersAPIOption {
	return func(c *UsersAPIController) {
		c.errorHandler = h
	}
}

// NewUsersAPIController creates a default api controller
func NewUsersAPIController(s UsersAPIServicer, opts ...UsersAPIOption) *UsersAPIController {
	controller := &UsersAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the UsersAPIController
func (c *UsersAPIController) Routes() Routes {
	return Routes{
		"GetUser": Route{
			"GetUser",
			strings.ToUpper("Get"),
			"/api/v2/users/{userId}",
			c.GetUser,
		},
		"UpdateUser": Route{
			"UpdateUser",
			strings.ToUpper("Put"),
			"/api/v2/users/{userId}",
			c.UpdateUser,
		},
		"DeleteUser": Route{
			"DeleteUser",
			strings.ToUpper("Delete"),
			"/api/v2/users/{userId}",
			c.DeleteUser,
		},
		"GetUserProfileImageSmall": Route{
			"GetUserProfileImageSmall",
			strings.ToUpper("Get"),
			"/api/v2/users/{userId}/profile_picture_small",
			c.GetUserProfileImageSmall,
		},
		"GetUserProfileImage": Route{
			"GetUserProfileImage",
			strings.ToUpper("Get"),
			"/api/v2/users/{userId}/profile_picture",
			c.GetUserProfileImage,
		},
		"GetUserXp": Route{
			"GetUserXp",
			strings.ToUpper("Get"),
			"/api/v2/users/{userId}/xp",
			c.GetUserXp,
		},
		"GetUserAchievements": Route{
			"GetUserAchievements",
			strings.ToUpper("Get"),
			"/api/v2/users/{userId}/achievements",
			c.GetUserAchievements,
		},
		"ClaimUserAchievement": Route{
			"ClaimUserAchievement",
			strings.ToUpper("Post"),
			"/api/v2/users/{userId}/achievements/{achievementId}",
			c.ClaimUserAchievement,
		},
	}
}

// OrderedRoutes returns all the api routes in a deterministic order for the UsersAPIController
func (c *UsersAPIController) OrderedRoutes() []Route {
	return []Route{
		Route{
			"GetUser",
			strings.ToUpper("Get"),
			"/api/v2/users/{userId}",
			c.GetUser,
		},
		Route{
			"UpdateUser",
			strings.ToUpper("Put"),
			"/api/v2/users/{userId}",
			c.UpdateUser,
		},
		Route{
			"DeleteUser",
			strings.ToUpper("Delete"),
			"/api/v2/users/{userId}",
			c.DeleteUser,
		},
		Route{
			"GetUserProfileImageSmall",
			strings.ToUpper("Get"),
			"/api/v2/users/{userId}/profile_picture_small",
			c.GetUserProfileImageSmall,
		},
		Route{
			"GetUserProfileImage",
			strings.ToUpper("Get"),
			"/api/v2/users/{userId}/profile_picture",
			c.GetUserProfileImage,
		},
		Route{
			"GetUserXp",
			strings.ToUpper("Get"),
			"/api/v2/users/{userId}/xp",
			c.GetUserXp,
		},
		Route{
			"GetUserAchievements",
			strings.ToUpper("Get"),
			"/api/v2/users/{userId}/achievements",
			c.GetUserAchievements,
		},
		Route{
			"ClaimUserAchievement",
			strings.ToUpper("Post"),
			"/api/v2/users/{userId}/achievements/{achievementId}",
			c.ClaimUserAchievement,
		},
	}
}

// GetUser - Get a user by userId
func (c *UsersAPIController) GetUser(w http.ResponseWriter, r *http.Request) {
	userIdParam := chi.URLParam(r, "userId")
	if userIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"userId"}, nil)
		return
	}
	result, err := c.service.GetUser(r.Context(), userIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// UpdateUser - Update user information by userId
func (c *UsersAPIController) UpdateUser(w http.ResponseWriter, r *http.Request) {
	userIdParam := chi.URLParam(r, "userId")
	if userIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"userId"}, nil)
		return
	}
	var userUpdateDtoParam UserUpdateDto
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&userUpdateDtoParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertUserUpdateDtoRequired(userUpdateDtoParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertUserUpdateDtoConstraints(userUpdateDtoParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.UpdateUser(r.Context(), userIdParam, userUpdateDtoParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeleteUser - Delete a user by userId
func (c *UsersAPIController) DeleteUser(w http.ResponseWriter, r *http.Request) {
	userIdParam := chi.URLParam(r, "userId")
	if userIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"userId"}, nil)
		return
	}
	var bodyParam int32
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&bodyParam); err != nil && !errors.Is(err, io.EOF) {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	result, err := c.service.DeleteUser(r.Context(), userIdParam, bodyParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetUserProfileImageSmall - Get the small profile picture of a user by userId
func (c *UsersAPIController) GetUserProfileImageSmall(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	userIdParam := chi.URLParam(r, "userId")
	if userIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"userId"}, nil)
		return
	}
	var redirectParam bool
	if query.Has("redirect") {
		param, err := parseBoolParameter(
			query.Get("redirect"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "redirect", Err: err}, nil)
			return
		}

		redirectParam = param
	} else {
		var param bool = false
		redirectParam = param
	}
	result, err := c.service.GetUserProfileImageSmall(r.Context(), userIdParam, redirectParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetUserProfileImage - Get the profile picture of a user by userId
func (c *UsersAPIController) GetUserProfileImage(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	userIdParam := chi.URLParam(r, "userId")
	if userIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"userId"}, nil)
		return
	}
	var redirectParam bool
	if query.Has("redirect") {
		param, err := parseBoolParameter(
			query.Get("redirect"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "redirect", Err: err}, nil)
			return
		}

		redirectParam = param
	} else {
		var param bool = false
		redirectParam = param
	}
	result, err := c.service.GetUserProfileImage(r.Context(), userIdParam, redirectParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetUserXp - Get user's xp
func (c *UsersAPIController) GetUserXp(w http.ResponseWriter, r *http.Request) {
	userIdParam := chi.URLParam(r, "userId")
	if userIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"userId"}, nil)
		return
	}
	result, err := c.service.GetUserXp(r.Context(), userIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetUserAchievements - Get user's achievements
func (c *UsersAPIController) GetUserAchievements(w http.ResponseWriter, r *http.Request) {
	userIdParam := chi.URLParam(r, "userId")
	if userIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"userId"}, nil)
		return
	}
	result, err := c.service.GetUserAchievements(r.Context(), userIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// ClaimUserAchievement - Claim an achievement
func (c *UsersAPIController) ClaimUserAchievement(w http.ResponseWriter, r *http.Request) {
	userIdParam := chi.URLParam(r, "userId")
	if userIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"userId"}, nil)
		return
	}
	achievementIdParam, err := parseNumericParameter[int32](
		chi.URLParam(r, "achievementId"),
		WithRequire[int32](parseInt32),
		WithMinimum[int32](0),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Param: "achievementId", Err: err}, nil)
		return
	}
	result, err := c.service.ClaimUserAchievement(r.Context(), userIdParam, achievementIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}